requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-API==3.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.3
cohere==4.37
wikipedia==1.4.0
openai==1.6.1
pydub==0.25.1
gTTS==2.5.0
langchain==0.0.352
chromadb==0.4.21
SpeechRecognition==3.10.1
python-telegram-bot==20.7
googletrans-py39==4.0.2
